fix duplicated root folder
wahoo
