--- Social Credit Leaderboard/static/css/style.css
@@ -1,44 +1,0 @@
-* {
-    font-family: 'Roboto', sans-serif;
-    background-color: hsl(250, 24%, 19%);
-    color: #fff;
-    scrollbar-width: none; /*firefox support */    
-}
-
-/* ********************************** */
-/*        AI RESPONSE STYLING         */
-/* ********************************** */
-
-/* Defines the width and alignment of the OpenAi text */
-#openai-container {
-    width: 450px;
-    text-align: center;
-}
-
-.openai-text {
-    font-size: 40px;
-    color: rgb(255, 255, 255);
-    text-shadow: 0px 0px 1px #000, 0px 0px 2px #000, 0px 0px 3px #000, 0px 0px 4px #000, 0px 0px 5px #000, 0px 0px 6px #000, 0px 0px 7px #000;
-    /* space above each list container */
-    margin: 4px 0 0 0;
-    
-    /* These are necessary if I'm animating the letters */
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 600px; /* Ensure it doesn't exceed 600px */
-}
-
-/* We put all of the letterized letters into these words elements */
-.openai-word {
-    display: inline-block;
-    display: flex;
-    flex-wrap: wrap;
-}
-
-/* Applies to each individual letter */
-.openai-letter {
-    color: rgb(255, 255, 255);
-    font-weight: bold;
-    display: inline-block; /* allow transform animations */
-}
