--- Social Credit Leaderboard/templates/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <title>Twitch Chat Sentiment</title>
-    <link rel="icon" type="image/x-icon" href="images/PepperLogo1.png">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Roboto&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.min.js" integrity="sha512-bLT0Qm9VnAYZDflyKcBaQ2gg0hSYNQrJ8RilYldYQ1FxQYoCLtUjuuRuZo+fjqhx/qtq/1itJ0C2ejDxltZVFg==" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/3.0.4/socket.io.js" integrity="sha512-aMGMvNYu8Ue4G+fHa359jcPb1u+ytAF+P2SCb+PxrjCdO3n3ZTxJ30zuH39rimUggmTwmh2u7wvQsDTHESnmfQ==" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/jquery-textfill@0.6.0/source/jquery.textfill.min.js"></script>
-    <script type="module" src="static/js/customTTS.js" defer></script>
-</head>
-<body>
-    <div id="top5-text" class="openai-text">Temporary Top 5</div>
-    
-
-    <div id="bottom5-text" class="openai-text">Temporary Bottom 5</div>
-
-</body>
-</html>
