+++ __pycache__/voices_manager.cpython-39.pyc
