+++ static/css/style.css
@@ -5,41 +5,40 @@
     scrollbar-width: none; /*firefox support */    
 }
 
-/* Defines the width and alignment of the overall donation*/
-#donation-container {
-    width: 600px;
+/* ********************************** */
+/*        AI RESPONSE STYLING         */
+/* ********************************** */
+
+/* Defines the width and alignment of the OpenAi text */
+#openai-container {
+    width: 450px;
     text-align: center;
 }
 
-/* Flexbox that contains the username, bits, and filler text */
-.donation-text {
-    font-size: 31px;
+.openai-text {
+    font-size: 40px;
     color: rgb(255, 255, 255);
-    text-shadow: 0px 0px 1px #000, 0px 0px 2px #000, 0px 0px 3px #000, 0px 0px 4px #000, 0px 0px 5px #000;
+    text-shadow: 0px 0px 1px #000, 0px 0px 2px #000, 0px 0px 3px #000, 0px 0px 4px #000, 0px 0px 5px #000, 0px 0px 6px #000, 0px 0px 7px #000;
+    /* space above each list container */
+    margin: 4px 0 0 0;
+    
+    /* These are necessary if I'm animating the letters */
     display: flex;
-    justify-content: center;
+    flex-direction: column;
     align-items: center;
-    gap: 1px; /* gap between each of the elements in the flexbox */
+    width: 600px; /* Ensure it doesn't exceed 600px */
 }
 
-.username {
-    margin-right: 4px;
-}
-
-.username, .bits {
-    font-size: 35px;
-    text-shadow: 0px 0px 1px #000, 0px 0px 2px #000, 0px 0px 3px #000, 0px 0px 4px #000, 0px 0px 5px #000;
+/* We put all of the letterized letters into these words elements */
+.openai-word {
+    display: inline-block;
     display: flex;
+    flex-wrap: wrap;
 }
 
-/* Applies to each individual letter of the username and bits strings */
-.letter {
-    color: rgb(0, 255, 255);
-}
-
-.message-text {
-    font-size: 26px;
+/* Applies to each individual letter */
+.openai-letter {
     color: rgb(255, 255, 255);
-    text-shadow: 0px 0px 1px #000, 0px 0px 2px #000, 0px 0px 3px #000, 0px 0px 4px #000, 0px 0px 5px #000;
-    margin-top: 4px;
+    font-weight: bold;
+    display: inline-block; /* allow transform animations */
 }
