+++ templates/index.html
@@ -1,8 +1,8 @@
 <!DOCTYPE HTML>
 <html>
 <head>
-    <title>DougDoug Custom TTS</title>
-    <link rel="icon" href="/favicon.ico" type="image/x-icon">
+    <title>Twitch Chat Sentiment</title>
+    <link rel="icon" type="image/x-icon" href="images/PepperLogo1.png">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Roboto&display=swap" rel="stylesheet">
@@ -13,13 +13,10 @@
     <script type="module" src="static/js/customTTS.js" defer></script>
 </head>
 <body>
-    <div id="donation-container">
-        <div class="donation-text">
-            <div id="username" class="username"></div>
-            <div>scored x</div>
-            <div id="bits" class="bits"></div>
-        </div>
-        <div id="message-text" class="message-text"></div>
-    </div>
+    <div id="top5-text" class="openai-text">Temporary Top 5</div>
+    
+
+    <div id="bottom5-text" class="openai-text">Temporary Bottom 5</div>
+
 </body>
 </html>
